finish gross refactor, preventing circular imports
moving lots of things from wcwidth.wcwidth to sub-modules, _constants,
_wcswidth, _wcwidth, and _width, _constants, and escap_esequences,
notably so they can be shared, such as _wcswidth in new
wcwidth.text_sizing, and existing iter_sequences by wcwidth.textwrap
without any "late-binding import" to avoid circular import.
