+++ docs/unicode_version.rst
@@ -16,6 +16,21 @@ release files:
 ``emoji-variation-sequences-12.0.0.txt``
   *Date: 2019-01-15, 12:10:05 GMT*
 
+``emoji-variation-sequences-13.0.0.txt``
+  *Date: 2020-01-21, 07:15:05 GMT*
+
+``emoji-variation-sequences-14.0.0.txt``
+  *Date: 2021-06-08, 05:19:16 GMT*
+
+``emoji-variation-sequences-15.0.0.txt``
+  *Date: 2022-05-13, 21:54:24 GMT*
+
+``emoji-variation-sequences-15.1.0.txt``
+  *Date: 2023-02-01, 02:22:54 GMT*
+
+``emoji-variation-sequences-16.0.0.txt``
+  *Date: 2024-05-01, 21:25:24 GMT*
+
 ``emoji-variation-sequences-17.0.0.txt``
   *Date: 2025-01-30, 21:48:29 GMT*
