+++ tox.ini
@@ -8,7 +8,6 @@ pip_compile_command = pip-compile --resolver=backtracking --strip-extras --no-em
 [testenv]
 deps = -r requirements-tests39.txt
 commands = {envpython} -m pytest --cov-config={toxinidir}/tox.ini {posargs:\
-             -n auto \
              --verbose \
              --junit-xml=.tox/results.{envname}.xml \
              --durations=3 \
